// src/app/shared/models/professor.model.ts
export interface Professor {
    firstName: string;
    lastName: string;
    email: string;
    gender: string;
    cin: string;
    profilePicture: string;
    degree: string;
    fieldOfExpertise: string;
    yearsOfExperience: number;
    phoneNumber: string;
    dateOfBirth: Date;
    password: string;
    confirmPassword: string;
}